+++ src/main/java/frc/robot/commands/gpm/IntakeNote.java
@@ -3,7 +3,7 @@ package frc.robot.commands.gpm;
 import edu.wpi.first.math.filter.Debouncer;
 import edu.wpi.first.wpilibj2.command.Command;
 import frc.robot.constants.IntakeConstants;
-import frc.robot.subsystems.gpm_subsystem.Intake;
+import frc.robot.subsystems.gpm.Intake;
 
 public class IntakeNote extends Command {
