+++ src/main/java/frc/robot/subsystems/gpm/Outtake.java
@@ -1,0 +1,5 @@
+package frc.robot.subsystems.gpm;
+
+public class Outtake {
+    
+}
